refactor: extract applyWorldStateThreadUpdates out of narrativeStore
The function lived inside narrativeStore.ts as a private async helper but
did pure cross-store orchestration: read sessionStore/worldStore/character
Store, build a thread/relationship/major-event update payload, optionally
hit /api/narrative/validate-event-significance, then call
worldStore.updateWorldState. It had a single call site inside addSegment
and didn't need to be co-located with the Zustand store.

Moved to src/lib/narrative/applyWorldStateThreadUpdates.ts alongside the
other narrative processors (itemAcquisitionProcessor, storyCheckpointPayload,
etc.). narrativeStore.ts is now 224 lines smaller (1465 -> 1241).
